fixed int nbins problem

+++ Simulation.py
@@ -573,7 +573,7 @@ kb = 1
 skin = 0.3
 
 #binwidth = Lz / nbins #add Rzwall and Lzwall to Lz
-nbins = Lz / binwidth + 1
+nbins = int(Lz / binwidth) + 1
 Volbin = Lx * Ly * binwidth
 totalLength = binwidth * nbins
